Fixed changes from commit conversations
